+++ src/screens/settingScreen/Settings.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import SettingsNavigator from './SettingsNavigator';
+
+export default function Settings() {
+  return <SettingsNavigator />;
+}
